Object-Oriented Update - Project, Directory, and Image Classes

+++ includes/classes/Project.php
@@ -1,80 +1,14 @@
 <?php
 
 
-class Link {
-  public $id;
-  public $text;
-  public $url;
-  public $project_id;
-  public $is_primary;
-
-  function __construct(
-    $mysqli,
-    $in_id=null,
-    $in_text=null,
-    $in_url=null,
-    $in_project_id=null,
-    $in_is_primary=null
-  ) {
-    if (!empty($in_id)) {         $this->id = $in_id; }
-    if (!empty($in_text)) {       $this->text = $in_text; }
-    if (!empty($in_url)) {        $this->url = $in_url; }
-    if (!empty($in_project_id)) { $this->project_id = $in_project_id; }
-    if (isset($in_is_primary)) {  $this->is_primary = $in_is_primary; }
-  }
-
-  function insertDB($mysqli) {
-    getResults(
-      $mysqli,
-      "INSERT INTO project_links (link_text, url, project_id, is_primary_link) VALUES (?, ?, ?, ?)",
-      "ssii",
-      array($this->text, $this->url, $this->project_id, $this->is_primary)
-    );
-  }
-
-  function deleteDB($mysqli) {
-    getResults(
-      $mysqli,
-      "DELETE FROM project_links WHERE id = ?",
-      "i",
-      array($this->id)
-    );
-  }
-
-  function updateDB($mysqli) {
-    $result = getResults(
-      $mysqli,
-      "SELECT * FROM project_links WHERE id = ?",
-      "i",
-      array($this->id)
-    );
-    $row = $result->fetch_assoc();
-
-    if ($row["link_text"] !== $this->text) { $this->updateTextDB($mysqli); }
-    if ($row["url"] !== $this->url) { $this->updateUrlDB($mysqli); }
-  }
-
-  private function updateTextDB($mysqli) {
-    getResults(
-      $mysqli,
-      "UPDATE project_links SET link_text=? WHERE id=?",
-      "si",
-      array($this->text, $this->id)
-    );
-  }
-
-  private function updateUrlDB($mysqli) {
-    getResults(
-      $mysqli,
-      "UPDATE project_links SET url=? WHERE id=?",
-      "si",
-      array($this->url, $this->id)
-    );
-  }
-}
+require_once (__DIR__ .'./Link.php');
+require_once (__DIR__ .'./Directory.php');
+require_once (__DIR__ .'./Image.php');
 
 
 class Project {
+  private $path = "./img/projects/";
+
   public $id;
   public $title;
   public $directory;
@@ -102,13 +36,17 @@ class Project {
 
     if (!empty($in_id)) {          $this->id = $in_id; }
     if (!empty($in_title)) {       $this->title = $in_title; }
-    if (!empty($in_directory)) {   $this->directory = $in_directory; }
-    if (!empty($in_image)) {       $this->image = $in_image; }
     if (!empty($in_blurb)) {       $this->blurb = $in_blurb; }
     if (!empty($in_description)) { $this->description = $in_description; }
     if (!empty($in_date)) {        $this->date = $in_date; }
     if (isset($in_featured)) {     $this->featured = $in_featured; }
     if (!empty($in_author_id)) {   $this->author_id = $in_author_id; }
+
+    if (!empty($in_directory)) {
+      $this->directory = new Directory($this->path, $in_directory);
+    }
+
+    $this->image = new Image($this->path, $in_image);
 
     $primary_link_object = null;
     $other_links_array = array();
@@ -148,7 +86,15 @@ class Project {
     $this->other_links = $other_links_array;
   }
 
-  function insertDB($mysqli) {
+  function create($mysqli, $img_file) {
+
+    // Configure Server Directory
+    $ok = $this->directory->create();   // create directory
+    if (!$ok) return false;
+    $ok = $this->image->upload($img_file);  // upload image
+    if (!$ok) return false;
+
+    // Insert Self into Database
     getResults(
       $mysqli,
       "INSERT INTO projects VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)",
@@ -156,8 +102,8 @@ class Project {
       array(
         $this->id,
         $this->title,
-        $this->directory,
-        $this->image,
+        $this->directory->name,
+        $this->image->name,
         $this->blurb,
         $this->description,
         $this->date,
@@ -173,14 +119,18 @@ class Project {
     }
   }
 
-  function deleteDB($mysqli) {
-    // Delete Links from Database
+  function delete($mysqli) {
+    // Delete image and directory from server (directory will remove image)
+    $ok = $this->directory->delete();
+    if (!$ok) return false;
+
+    // Delete links from database
     $this->primary_link->deleteDB($mysqli);
     foreach ($this->other_links as $link) {
       $link->deleteDB($mysqli);
     }
 
-    // Delete self from Database
+    // Delete self from database
     getResults(
       $mysqli,
       "DELETE FROM projects WHERE ID = ?",
@@ -189,24 +139,52 @@ class Project {
     );
   }
 
-  function updateDB($mysqli) {
+  function update($mysqli, $new_img_file=null) {
+    // grab current details database
     $result = getResults(
       $mysqli,
       "SELECT * FROM projects WHERE ID = ?",
       "i",
       array($this->id)
     );
-
     $row = $result->fetch_assoc();
 
-    if ($row["title"] !== $this->title) {             $this->updateTitleDB($mysqli); }
-    if ($row["directory"] !== $this->directory) {     $this->updateDirectoryDB($mysqli); }
-    if ($row["image"] !== $this->image) {             $this->updateImageDB($mysqli); }
-    if ($row["blurb"] !== $this->blurb) {             $this->updateBlurbDB($mysqli); }
-    if ($row["description"] !== $this->description) { $this->updateDescriptionDB($mysqli); }
-    if ($row["date"] !== $this->date) {               $this->updateDateDB($mysqli); }
-    if ($row["featured"] !== $this->featured) {       $this->updateFeaturedDB($mysqli); }
-    if ($row["author_id"] !== $this->author_id) {     $this->updateAuthorDB($mysqli); }
+    // compare self to database
+    if ($row["directory"] !== $this->directory->name) { $this->updateDirectory($mysqli); }
+    if (!empty($new_img_file)) {                        $this->updateImage($mysqli, $new_img_file); }
+
+    if ($row["title"] !== $this->title) {               $this->updateTitleDB($mysqli); }
+    if ($row["blurb"] !== $this->blurb) {               $this->updateBlurbDB($mysqli); }
+    if ($row["description"] !== $this->description) {   $this->updateDescriptionDB($mysqli); }
+    if ($row["date"] !== $this->date) {                 $this->updateDateDB($mysqli); }
+    if ($row["featured"] !== $this->featured) {         $this->updateFeaturedDB($mysqli); }
+    if ($row["author_id"] !== $this->author_id) {       $this->updateAuthorDB($mysqli); }
+  }
+
+  private function updateDirectory($mysqli) {
+    // update server
+    $ok = $this->directory->rename($this->directory->name);
+    if (!$ok) return false;
+    // update database
+    getResults(
+      $mysqli,
+      "UPDATE projects SET directory=? WHERE ID=?",
+      "si",
+      array($this->directory->name, $this->id)
+    );
+  }
+
+  private function updateImage($mysqli, $new_img_file) {
+    // update server
+    $ok = $this->image->upload($new_img_file);
+    if (!$ok) return false;
+    // update database
+    getResults(
+      $mysqli,
+      "UPDATE projects SET image=? WHERE ID=?",
+      "si",
+      array($this->image->name, $this->id)
+    );
   }
 
   private function updateTitleDB($mysqli) {
@@ -215,24 +193,6 @@ class Project {
       "UPDATE projects SET title=? WHERE ID=?",
       "si",
       array($this->title, $this->id)
-    );
-  }
-
-  private function updateDirectoryDB($mysqli) {
-    getResults(
-      $mysqli,
-      "UPDATE projects SET directory=? WHERE ID=?",
-      "si",
-      array($this->directory, $this->id)
-    );
-  }
-
-  private function updateImageDB($mysqli) {
-    getResults(
-      $mysqli,
-      "UPDATE projects SET image=? WHERE ID=?",
-      "si",
-      array($this->image, $this->id)
     );
   }
